Fixing new feature

+++ master-ugr.calculator.lib/calculator.lib/NumberAttributter.cs
@@ -30,7 +30,7 @@ namespace calculator.lib
             return number % 2 != 0;
         }
 
-        public static double SquareRoot(int number)
+        public static double SquareRoot(double number)
         {
             return Math.Sqrt(number);
         }
